Fix hardcoding of local IP

+++ uninstall.sh
@@ -1,8 +1,11 @@
 #!/bin/bash
-# Version: 0.1a
+# Version: 0.1.1
 # Date: 2023-09-16
 # Dependencies: Assumes Ubuntu or Debian-based system with apt package manager.
 # Description: Uninstall and revert services and packages installed by the install script.
+
+# List of packages installed by the installation script
+PACKAGES_INSTALLED=("docker" "docker-engine" "docker.io" "containerd" "runc" "nginx" "netdata")
 
 # Check if running with sudo
 if [ "$EUID" -eq 0 ]; then
@@ -27,22 +30,27 @@ is_installed() {
 # Function to revert changes
 revert_packages() {
   # Restore package list
-  sudo dpkg --clear-selections
-  sudo dpkg --set-selections < "${HOME}/system_state_tracking/package_list_before.txt"
+  for pkg in "${PACKAGES_INSTALLED[@]}"; do
+    sudo dpkg --set-selections < "${HOME}/system_state_tracking/package_list_before.txt"
+  done
   sudo apt-get dselect-upgrade -y
 }
 
 # Delete any new files
 revert_files() {
-comm -13 "${HOME}/system_state_tracking/filesystem_before.txt" "${HOME}/system_state_tracking/filesystem_after.txt" | while read -r line; do
-  if [[ "$line" == ${HOME}* ]]; then  # only delete files in the user's home directory to be safe
-    sudo rm -rf "$line"
-  fi
-done
+  comm -13 "${HOME}/system_state_tracking/filesystem_before.txt" "${HOME}/system_state_tracking/filesystem_after.txt" | while read -r line; do
+    if [[ "$line" == ${HOME}* ]]; then  # only delete files in the user's home directory to be safe
+      sudo rm -rf "$line"
+    fi
+  done
 }
 
 # Uninstall Docker
-sudo apt-get remove docker docker-engine docker.io containerd runc
+for pkg in "${PACKAGES_INSTALLED[@]}"; do
+  if [[ "$pkg" == "docker" || "$pkg" == "docker-engine" || "$pkg" == "docker.io" || "$pkg" == "containerd" || "$pkg" == "runc" ]]; then
+    sudo apt-get remove "$pkg"
+  fi
+done
 
 # Remove Docker Compose
 sudo rm /usr/local/bin/docker-compose
@@ -64,7 +72,9 @@ if is_installed "nginx"; then
 fi
 
 # Remove Netdata
-sudo apt remove -y netdata
+if is_installed "netdata"; then
+  sudo apt remove -y netdata
+fi
 
 # Restore the previous package list and delete new files
 revert_packages
